Fixed deprecated Ruby code.
exists? is deprecated as of Ruby 2.2, exist? (without the s and the end) replaces it apparently. (As of writing Ruby version is 3.4)

I added links to install Ruby and webrick.

I am printing http://127.0.0.1:8000 into the console, because it makes it much easier to use. You can copy-paste or click the link to open it.